add category

+++ app/Http/Controllers/Admin/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use Illuminate\Support\Str;
 use App\Model\Post;
+use App\Model\Category;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -22,13 +23,26 @@ class PostController extends Controller
     }
 
     /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function indexUser()
+    {
+        $posts = Post::where('user_id', Auth::user()->id)->orderBy('created_at', 'desc')->paginate(10);
+
+        return view('admin.posts.index', ['posts' => $posts]);
+    }
+
+    /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
     {
-        return view('admin.posts.create');
+        $categories = Category::all();
+        return view('admin.posts.create', ['categories' => $categories]);
     }
 
     /**
@@ -48,21 +62,12 @@ class PostController extends Controller
         $data = $request->all();
         $data['user_id'] = Auth::user()->id;
 
-        $slug = Str::slug($data['title'], '-');
-        $postPresente = Post::where('slug', $slug)->first();
 
-
-        $counter = 0;
-        while ($postPresente) {
-            $slug = $slug . '-' . $counter;
-            $postPresente = Post::where('slug', $slug)->first();
-            $counter++;
-        }
 
         $newPost = new Post();
 
         $newPost->fill($data);
-        $newPost->slug = $slug;
+        $newPost->slug = $newPost->createSlug($data['title']);
         $newPost->save();
 
         return redirect()->route('admin.posts.show', $newPost->slug);
@@ -76,7 +81,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return view('admin.posts.show', ['post' => $post]);
+        $data = ['post' => $post];
+        return view('admin.posts.show', $data);
     }
 
     /**
@@ -87,7 +93,12 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        return view('admin.posts.edit', ['post' => $post]);
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+        $categories = Category::all();
+
+        return view('admin.posts.edit', ['post' => $post, 'categories' => $categories]);
     }
 
     /**
@@ -102,12 +113,26 @@ class PostController extends Controller
         $validateData = $request->validate([
             'title' => 'required|max:255',
             'content' => 'required',
+            'category_id' => 'exsist:App\Model\Category,id',
         ]);
         $data = $request->all();
-        $updated = $post->update($data);
-        if (!$updated) {
-            dd('aggiornamento non riuscito');
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
         }
+
+        if ($data['title'] != $post->title) {
+            $post->title = $data['title'];
+            $post->slug = $post->createSlug($data['title']);
+        }
+        if ($data['content'] != $post->content) {
+            $post->content = $data['content'];
+        }
+        if ($data['category_id'] != $post->category_id) {
+            $post->category_id = $data['category_id'];
+        }
+
+        $post->update();
+
         return redirect()
             ->route('admin.posts.show', $post);
     }
@@ -120,10 +145,12 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+
         $post->delete();
 
-        return redirect()
-            ->route('admin.posts.index')
-            ->with('status', "Post id $post->id deleted");
+        return redirect()->route('admin.posts.index')->with('status', "Post id $post->id deleted");
     }
 }
